requirements.txt
flask
gunicorn
tensorflow
numpy
pillow